navigator: extract AppLayout view; lint container/presenter split
- views/AppLayout.tsx: move App + 3 layout containers out of create-app
- create-app.tsx becomes pure DI/composition root (389 -> 159 LOC); drop
  render-prop slot ceremony for views
- eslint: ban stores/controllers value imports from components/
  (allowTypeImports preserves existing type-only imports)
- components/RouteStack: drop dead useMeasure refs
